Refactor upload benchmark code.
