+++ benchmark/upload/CMakeLists.txt
@@ -3,7 +3,10 @@ add_executable(upload_benchmark upload_benchmark.cpp
                                 grpc_dummy_satellite.cpp
                                 stream_dummy_satellite.cpp
                                 stream_session.cpp
-                                upload_benchmark_report.cpp)
+                                upload_benchmark.cpp
+                                upload_benchmark_report.cpp
+                                tracer.cpp
+                                main.cpp)
 target_link_libraries(upload_benchmark lightstep_tracer
                                        ${LIGHTSTEP_LINK_LIBRARIES}
                                        $<TARGET_OBJECTS:upload_benchmark_configuration>)
