+++ benchmark/upload/tracer.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "configuration-proto/tracer_configuration.pb.h"
+
+#include <opentracing/tracer.h>
+
+namespace lightstep {
+std::shared_ptr<opentracing::Tracer> MakeTracer(
+    const tracer_configuration::TracerConfiguration& configuration);
+}  // namespace lightstep
