+++ src/lightstep_tracer_factory.cpp
@@ -4,6 +4,9 @@
 #include "configuration-proto/tracer_configuration.pb.h"
 
 namespace lightstep {
+//------------------------------------------------------------------------------
+// MakeTracer
+//------------------------------------------------------------------------------
 opentracing::expected<std::shared_ptr<opentracing::Tracer>>
 LightStepTracerFactory::MakeTracer(const char* configuration,
                                    std::string& error_message) const
